✏️  : 세미콜론 삭제
