--- hooks/useInterval.ts
@@ -1,22 +1,0 @@
-import { useEffect, useRef } from "react";
-
-export const useInterval = (callback: () => void, delay: number | null) => {
-  const savedCallback = useRef<() => void>(() => {});
-
-  useEffect(() => {
-    savedCallback.current = callback;
-  }, [callback]);
-
-  useEffect(() => {
-    function tick() {
-       if (typeof savedCallback.current === "function") {
-    savedCallback.current();
-  }
-    }
-
-    if (delay !== null) {
-      const id = setInterval(tick, delay);
-      return () => clearInterval(id);
-    }
-  }, [delay]);
-}
